--- config/buildkitd.toml
@@ -1,3 +1,0 @@
-[registry."89.167.60.202:5000"]
-  http = true
-  insecure = true
