--- config/clickhouse/ipv4-only.xml
@@ -1,5 +1,0 @@
-<clickhouse>
-    <listen_host>0.0.0.0</listen_host>
-    <http_port>8123</http_port>
-    <tcp_port>9000</tcp_port>
-</clickhouse>
