--- config/clickhouse/low-resources.xml
@@ -1,3 +1,0 @@
-<clickhouse>
-    <mark_cache_size>524288000</mark_cache_size>
-</clickhouse>
