--- config/discourse/install-ssl
@@ -1,2 +1,0 @@
-#!/bin/bash
-echo "Skipping SSL setup (handled by kamal-proxy)"
