--- config/plausible/clickhouse/ipv4-only.xml
@@ -1,3 +1,0 @@
-<clickhouse>
-    <listen_host>0.0.0.0</listen_host>
-</clickhouse>
