--- config/prometheus/prometheus.yml
@@ -1,3 +1,0 @@
-global:
-  scrape_interval: 60s
-  evaluation_interval: 60s
